requirements.txt
seafileapi2